changed picarx_improved dir_cal_value to float

+++ picarx/picarx_improved.py
@@ -57,7 +57,7 @@ class Picarx(object):
         self.camera_servo_pin1 = Servo(PWM(servo_pins[0]))
         self.camera_servo_pin2 = Servo(PWM(servo_pins[1]))   
         self.dir_servo_pin = Servo(PWM(servo_pins[2])) 
-        self.dir_cal_value = int(self.config_flie.get("picarx_dir_servo", default_value=0))
+        self.dir_cal_value = float(self.config_flie.get("picarx_dir_servo", default_value=0))
         self.cam_cal_value_1 = int(self.config_flie.get("picarx_cam_servo1", default_value=0))
         self.cam_cal_value_2 = int(self.config_flie.get("picarx_cam_servo2", default_value=0))
         self.dir_servo_pin.angle(self.dir_cal_value)
